Move default ajax function out of index so that it can be tested.
