+++ src/ajax.js
@@ -1,0 +1,30 @@
+// https://gist.github.com/Xeoncross/7663273
+function ajax(url, options, callback, data, cache) {
+  // Must encode data
+  if(data && typeof data === 'object') {
+    var y = '', e = encodeURIComponent;
+    for (var m in data) {
+      y += '&' + e(m) + '=' + e(data[m]);
+    }
+    data = y.slice(1) + (!cache ? '&_t=' + new Date : '');
+  }
+
+  try {
+    var x = new (XMLHttpRequest || ActiveXObject)('MSXML2.XMLHTTP.3.0');
+    x.open(data ? 'POST' : 'GET', url, 1);
+    if (!options.crossDomain) {
+      x.setRequestHeader('X-Requested-With', 'XMLHttpRequest');
+    }
+    if (data) {
+      x.setRequestHeader('Content-type', 'application/x-www-form-urlencoded');
+    }
+    x.onreadystatechange = function() {
+      x.readyState > 3 && callback && callback(x.responseText, x);
+    };
+    x.send(data);
+  } catch (e) {
+    window.console && console.log(e);
+  }
+}
+
+export default ajax;
