+++ src/index.js
@@ -1,64 +1,5 @@
 import * as utils from './utils';
-
-// https://gist.github.com/Xeoncross/7663273
-function ajax(url, options, callback, data, cache) {
-  // Must encode data
-  if(data && typeof data === 'object') {
-    var y = '', e = encodeURIComponent;
-    for (var m in data) {
-      y += '&' + e(m) + '=' + e(data[m]);
-    }
-    data = y.slice(1) + (!cache ? '&_t=' + new Date : '');
-  }
-
-  try {
-    var x = new (XMLHttpRequest || ActiveXObject)('MSXML2.XMLHTTP.3.0');
-    x.open(data ? 'POST' : 'GET', url, 1);
-    if (!options.crossDomain) {
-      x.setRequestHeader('X-Requested-With', 'XMLHttpRequest');
-    }
-    if (data) {
-      x.setRequestHeader('Content-type', 'application/x-www-form-urlencoded');
-    }
-    x.onreadystatechange = function() {
-      x.readyState > 3 && callback && callback(x.responseText, x);
-    };
-    x.send(data);
-  } catch (e) {
-    window.console && console.log(e);
-  }
-};
-
-// ajax.uriEncode = function(o) {
-//     var x, y = '', e = encodeURIComponent;
-//     for (x in o) y += '&' + e(x) + '=' + e(o[x]);
-//     return y.slice(1);
-// };
-//
-// ajax.collect = (a, f) {
-//     var n = [];
-//     for (var i = 0; i < a.length; i++) {
-//         var v = f(a[i]);
-//         if (v != null) n.push(v);
-//     }
-//     return n;
-// };
-//
-// ajax.serialize = function (f) {
-//     function g(n) {
-//         return f.getElementsByTagName(n);
-//     };
-//     var nv = function (e) {
-//         if (e.name) return encodeURIComponent(e.name) + '=' + encodeURIComponent(e.value);
-//     };
-//     var i = collect(g('input'), function (i) {
-//         if ((i.type != 'radio' && i.type != 'checkbox') || i.checked) return nv(i);
-//     });
-//     var s = collect(g('select'), nv);
-//     var t = collect(g('textarea'), nv);
-//     return i.concat(s).concat(t).join('&');
-// };
-//
+import ajax from './ajax'
 
 function getDefaults() {
   return {
